refactor: home the daemon-inventory row schemas in @kolu/padi/surface, not @kolu/terminal-workspace
srid rejected the first placement. RunningKaval/RunningPadi/HostDaemonInventorySchema were
put on @kolu/terminal-workspace/schema on a MISREAD of the seal — I wrote "the seal forbids
either surface package importing the other", but the seal only forbids PADI importing kolu
packages. The reverse (kolu-common/client/server importing terminal/daemon vocabulary from
@kolu/padi/surface) is the established direction since W1.M (surfacesWithPadi, contract,
the dialogs all do it directly, no shim).

@kolu/terminal-workspace is TERMINAL-AWARENESS vocabulary (sensors · fold · agentProjection ·
schema-of-terminals). A kaval gate pid is a DAEMON-domain fact — putting it there made the
awareness library know what a daemon is, a boundary violation (location is structure). So:

- Moved the three schemas + types to @kolu/padi/surface (padi owns the daemon domain),
  beside the other padiSurface schemas; reverted the @kolu/terminal-workspace diff entirely.
- kolu-common/surface's daemonInventory cell imports HostDaemonInventorySchema from
  @kolu/padi/surface (the kolu-common → @kolu/padi edge the package already uses); the row
  TYPES are re-exported from there so the client dialogs' imports are unchanged.
- padi/hostInventory.ts reads the rows from ./surface (local).
- One consequence of the correct import edge: useSessionRestore.test.ts's @kolu/padi/surface
  mock now spreads importOriginal so HostDaemonInventorySchema stays present.

No import-direction wall (@kolu/padi/surface imports nothing from kolu — acyclic). just check
+ the seal test + all touched unit suites green.
