+++ packages/terminal-workspace/src/schema.ts
@@ -262,92 +262,6 @@ export const ProcessRssSchema = z.discriminatedUnion("status", [
 ]);
 export type ProcessRss = z.infer<typeof ProcessRssSchema>;
 
-// ── Host-daemon inventory rows (the Kaval/Padi dialogs' leak diagnostic) ────
-//
-// One running kaval / padi the host-daemon scan enumerated — the read-only
-// diagnostic rows the Kaval + Padi info dialogs list so a LEAKED daemon (a
-// pre-upgrade kaval, a second padi at another state-root) is visible AT A GLANCE.
-// (srid hit this dogfooding W2.2: a leaked pre-W2.2 kaval was invisible in the UI
-// — only a `kaval-tui: more than one kaval daemon is running` CLI error surfaced
-// it.) Server-authored, read-only enumeration: scan the runtime dir, read each gate
-// pid, best-effort probe status — it NEVER kills/reaps/touches a daemon.
-//
-// These live on this browser-safe shared-vocab leaf (beside {@link ProcessRssSchema})
-// because BOTH sides of the padi seal compose them: `@kolu/padi/surface` serves the
-// bound host's own scan as its `hostInventory` member, and `kolu-common/surface`
-// carries kolu-server's local-machine scan on its `daemonInventory` cell. One scan
-// implementation (in `@kolu/padi`), one wire shape here — no lockstep copy held
-// together by a comment, and the seal forbids either surface package importing the
-// other.
-//
-// Honesty (#1034): every field the probe couldn't read is an honest `null` (rendered
-// "—"), never a fabricated zero/version.
-
-export const RunningKavalSchema = z.object({
-  /** The rendezvous socket path — the pasteable `--socket` value. */
-  socket: z.string(),
-  /** Discovery's human label ("standalone kaval" | "kolu @ <state-root>" |
-   *  "kolu-server on port <port>"), decided at discovery's matching branch. */
-  label: z.string(),
-  /** The structural kind: `stateRoot` (a padi's kaval — carries a state-root
-   *  manifest, incl. an ADOPTED legacy-address kaval), `port` (an UN-adopted legacy
-   *  `kaval-<port>/` with NO manifest — a genuine stray/leak), `standalone`, or
-   *  `unknown`. */
-  kind: z.enum(["stateRoot", "port", "standalone", "unknown"]),
-  /** The gate-holder pid (`kaval.pid`), or null if unreadable. */
-  gatePid: z.number().int().nullable(),
-  /** Live terminal count from a best-effort `terminal.list` probe, or null when the
-   *  probe failed / the daemon didn't answer (never a fake 0). */
-  terminalCount: z.number().int().nullable(),
-  /** The kaval's build commit (`navigableCommit`) from a best-effort `system.version`
-   *  probe, or null when unreadable. */
-  buildCommit: z.string().nullable(),
-  /** The pty-host contract version from the probe, or null when unreadable. */
-  contractVersion: z.string().nullable(),
-  /** Whether the scanning host's kolu ACTIVELY owns this kaval ("in use by kolu"), and —
-   *  when it does — whether it sits at the pre-padi LEGACY `kaval-<port>/` address (padi
-   *  ADOPTED a live pre-W2.2 kaval on upgrade rather than leaking it — a KNOWN converging
-   *  state, not a leak, until the next recycle spawns it at the digest address).
-   *
-   *  A discriminated pair, NOT two independent booleans: `atLegacyAddress` exists ONLY on
-   *  the `active` arm, so the nonsense "legacy-but-not-owned" state is UNREPRESENTABLE
-   *  (P4). Only the host serving its OWN `hostInventory` marks `active` — a local-machine
-   *  scan under a remote binding is always `{ active: false }` (kolu is bound elsewhere). */
-  held: z.discriminatedUnion("active", [
-    z.object({ active: z.literal(false) }),
-    z.object({ active: z.literal(true), atLegacyAddress: z.boolean() }),
-  ]),
-});
-export type RunningKaval = z.infer<typeof RunningKavalSchema>;
-
-export const RunningPadiSchema = z.object({
-  /** padi's rendezvous socket path. */
-  socket: z.string(),
-  /** padi's state-root (from the digest→root manifest), or null if unreadable. */
-  stateRoot: z.string().nullable(),
-  /** The gate-holder pid (`padi.pid`), or null if unreadable. */
-  gatePid: z.number().int().nullable(),
-  /** True iff this is the padi the scanning host's kolu owns ("in use by kolu"). The
-   *  active padi's contract version + build commit do NOT ride this row — padi cannot
-   *  probe a foreign padi, so every non-active row would carry nulls; the one bound
-   *  padi's identity is published once on `daemonInventory.boundPadi` (the honest
-   *  fresh-each-tick live read that also works over ssh). */
-  active: z.boolean(),
-});
-export type RunningPadi = z.infer<typeof RunningPadiSchema>;
-
-/** One host's daemon inventory — every running kaval + padi on a single machine.
- *  Homed on this browser-safe leaf beside the two row shapes it wraps so BOTH sides
- *  of the padi seal read ONE declaration: `@kolu/padi/surface`'s `hostInventory`
- *  member (the bound host's own scan) and `kolu-common/surface`'s
- *  `daemonInventory.localScan` (kolu-server's local-machine scan) compose this one
- *  container, not two lockstep copies. The scanner returns this neutral name. */
-export const HostDaemonInventorySchema = z.object({
-  kavals: z.array(RunningKavalSchema),
-  padis: z.array(RunningPadiSchema),
-});
-export type HostDaemonInventory = z.infer<typeof HostDaemonInventorySchema>;
-
 // ── Live-output cadence ────────────────────────────────────────────────────
 
 /** Output quiet-period before a terminal reads as static again — the ONE cadence
